Move out mock_io
